Fixed bugs, did some other things.

+++ Source/SimulationEngine/SimulationEngineModes/SimulationEngineSingleThread.h
@@ -24,9 +24,9 @@ public:
 
     static void produce_food    (EngineDataContainer * dc, SimulationParameters * sp, Organism *organism, lehmer64 &gen);
 
-    static void produce_food_simplified(EngineDataContainer * dc, SimulationParameters * sp, Organism *organism, lehmer64 &gen, int multiplier);
+    static void produce_food_simplified(EngineDataContainer * dc, SimulationParameters * sp, Organism *organism, lehmer64 &gen, float multiplier);
 
-    static void produce_food_complex(EngineDataContainer * dc, SimulationParameters * sp, Organism *organism, lehmer64 &gen, int multiplier);
+    static void produce_food_complex(EngineDataContainer * dc, SimulationParameters * sp, Organism *organism, lehmer64 &gen, float multiplier);
 
     static void eat_food        (EngineDataContainer * dc, SimulationParameters * sp, Organism *organism);
